Rename tequilapi metrics
