+++ src/app/bug-reporting/tequilapi-client-with-metrics.js
