+++ src/dependencies/mysterium-tequilapi.js
@@ -18,7 +18,7 @@
 // @flow
 import type {Container} from '../app/di'
 import {BugReporterMetrics} from '../app/bug-reporting/bug-reporter-metrics'
-import TequilapiClientWithMetrics from '../app/bug-reporting/tequilapi-metrics'
+import TequilapiClientWithMetrics from '../app/bug-reporting/tequilapi-client-with-metrics'
 import type {TequilapiClient} from '../libraries/mysterium-tequilapi/client'
 import tequilapiClientFactory from '../libraries/mysterium-tequilapi/client-factory'
